Account for "dirty" data when deciding weather or not to "skip" running the Laravel watcher for @entangle
