+++ js/Message.js
@@ -16,20 +16,21 @@ export default class {
         }
     }
 
-    shouldSkipWatcher() {
-        return this.updateQueue.length && this.updateQueue.every(update => update.skipWatcher)
-    }
-
     shouldSkipWatcherForDataKey(dataKey) {
+        // If the data is dirty, run the watcher.
+        if (this.response.effects.dirty.includes(dataKey)) return false
+
         let compareBeforeFirstDot = (subject, value) => {
             if (typeof subject !== 'string' || typeof value !== 'string') return false
 
             return subject.split('.')[0] === value.split('.')[0]
         }
 
+        // Otherwise see if there was a defered update for a data key.
+        // In that case, we want to skip running the Livewire watcher.
         return this.updateQueue
             .filter(update => compareBeforeFirstDot(update.name, dataKey))
-            .some(update => update.shouldSkipWatcher)
+            .some(update => update.skipWatcher)
     }
 
     storeResponse(payload) {
