+++ tests/Browser/Alpine/Entangle/DeferDataUpdates.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Tests\Browser\Alpine\Entangle;
+
+use Illuminate\Support\Facades\View;
+use Livewire\Component as BaseComponent;
+
+class DeferDataUpdates extends BaseComponent
+{
+    public $testing;
+
+    public function render()
+    {
+        return <<<'HTML'
+<div x-data="{ testing:  @entangle('testing').defer }">
+    <input type="text" x-model="testing" dusk="input">
+
+    <p>Alpine: <span dusk="output.alpine" x-text="testing"></span></p>
+
+    <p>Livewire: <span dusk="output.livewire">{{$testing}}</span></p>
+
+    <button wire:click.prevent="$refresh" dusk="submit">Submit</button>
+</div>
+HTML;
+    }
+}
